requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
influxdb-client==1.48.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
reactivex==4.0.4
requests==2.32.3
setuptools==80.4.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
# optional dev tools
black==24.8.0
ruff==0.5.6